creating new modules and breaking up metadata_generator class
