+++ src/bio_ontology_api.py
@@ -1,6 +1,5 @@
 # -*- coding: utf-8 -*-
 import requests
-from dotenv import load_dotenv
 import os
 from pathlib import Path
 import requests
