+++ src/tests/test_gcms_metadata_gen.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 # This script will serve as a test for the lipdomics metadata generation script.
 from datetime import datetime
-from src.metadata_generator import GCMSMetabolomicsMetadataGenerator
+from src.gcms_metadata_generator import GCMSMetabolomicsMetadataGenerator
 from dotenv import load_dotenv
 
 load_dotenv()
